requirements.txt
altgraph==0.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
demucs==4.0.1
dnspython==2.7.0
dora_search==0.1.12
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
fsspec==2025.2.0
future==0.18.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
julius==0.2.7
lameenc==1.8.1
llvmlite==0.43.0
macholib==1.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20240930
openunmix==1.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.15.0
sniffio==1.3.1
starlette==0.45.3
submitit==1.5.2
sympy==1.13.1
tiktoken==0.8.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
treetable==0.2.5
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
